+++ packages/iov-jsonrpc/src/index.ts
@@ -10,9 +10,7 @@ export {
 export {
   parseJsonRpcId,
   parseJsonRpcRequest,
-  parseJsonRpcResponse,
   parseJsonRpcResponse2,
-  parseJsonRpcError,
   parseJsonRpcErrorResponse,
   parseJsonRpcSuccessResponse,
 } from "./parse";