+++ packages/iov-jsonrpc/types/parse.d.ts
@@ -8,12 +8,8 @@ export declare function parseJsonRpcId(data: unknown): JsonRpcId | null;
 export declare function parseJsonRpcRequest(data: unknown): JsonRpcRequest;
 /** Throws if data is not a JsonRpcErrorResponse */
 export declare function parseJsonRpcErrorResponse(data: unknown): JsonRpcErrorResponse;
-/** @deprecated use parseJsonRpcErrorResponse */
-export declare function parseJsonRpcError(data: unknown): JsonRpcErrorResponse | undefined;
 /** Throws if data is not a JsonRpcSuccessResponse */
 export declare function parseJsonRpcSuccessResponse(data: unknown): JsonRpcSuccessResponse;
-/** @deprecated use parseJsonRpcSuccessResponse */
-export declare function parseJsonRpcResponse(data: unknown): JsonRpcSuccessResponse;
 /**
  * Returns a JsonRpcErrorResponse if input can be parsed as a JSON-RPC error. Otherwise parses
  * input as JsonRpcSuccessResponse. Throws if input is neither a valid error nor success response.